backend/requirements.txt
Django==5.2
djangorestframework==3.16
psycopg2-binary==2.9.10
python-dotenv==1.0.1
dj-database-url==2.3.0
drf-spectacular==0.29.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.6.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
